add btn btn-primary btn-xs to button view image

+++ module/Admin/src/Admin/Controller/MeasureController.php
@@ -15,20 +15,20 @@ class MeasureController extends AbstractActionController
 implements ConfigAwareInterface
 {
 	use AdminTablesTrait;
+
 	private $config;
-	
+
 	public function indexAction()
 	{
 		return new ViewModel(array(
-				'measures' => $this->getMeasureTable()->fetchAll(),
-				'config' => $this->config
-		));
+			'measures' => $this->getMeasureTable()->fetchAll(),
+			'config' => $this->config
+			));
 	}
 
 	public function addAction()
 	{
 		$form = $this->getServiceLocator()->get("Admin\Form\MeasureForm");
-		
 		$request = $this->getRequest();
 
 		if ($request->isPost()) {
@@ -36,67 +36,31 @@ implements ConfigAwareInterface
 			$measure = new Measure();
 			$form->setInputFilter($measure->getInputFilter());
 
-			$nonFile = $request->getPost()->toArray();
-			$file    = $this->params()->fromFiles('image');
-			$data = array_merge(
-					$nonFile, //POST
-					array('image'=> $file['name'])
-			);
+			$data = $request->getPost()->toArray();
 
 			$form->setData($data);
 
 			if ($form->isValid()) {
 
-			if(!file_exists($this->config['component']['measure']['image_path']))
-				mkdir($this->config['component']['measure']['image_path']);
-				
-			if(!empty($file['name'])) {
-				 $size = new Size($this->config['file_characteristics']['image']['size']);
-				 $extension= new Extension($this->config['file_characteristics']['image']['extension']);
+				$fileService = $this->getServiceLocator()->get('Admin\Service\FileService');
 
-				 $adapter = new HttpTransfer();
-				 $adapter->setValidators(array($size,$extension), $file['name']);
+				$fileService->setDestination($this->config['component']['measure']['image_path']);
+				$fileService->setSize($this->config['file_characteristics']['image']['size']);
+				$fileService->setExtension($this->config['file_characteristics']['image']['extension']);
 
-				 if (!$adapter->isValid()) {
-				 	dumpx("es valido");
-				 	$dataError = $adapter->getMessages();
+				$image = $fileService->copy($this->params()->fromFiles('image'));
+				$data['image'] = $image ? $image : "" ;
 
-				 	$error = array();
-				 	foreach($dataError as $key=>$row)
-				 		$error[] = $row;
+				$measure->exchangeArray($data);
+				$this->getBrandTable()->save($measure);
 
-				 	$form->setMessages(array('image'=>$error));
-				 }
-				 else {
-				 	$adapter->setDestination($this->config['component']['measure']['image_path']);
-
-					 	foreach ($adapter->getFileInfo() as $info) {
-					 		$file_name = time().'.'.pathinfo($info['name'], PATHINFO_EXTENSION) ;
-					 			
-					 		$adapter->addFilter('File\Rename',
-					 				array('target' => $adapter->getDestination().'/'.$file_name.'',
-					 						'overwrite' => true));
-
-					 		if ($adapter->receive($info['name'])) {
-
-					 			$measure->exchangeArray($form->getData());
-					 			$measure->setImage($file_name);
-					 			$this->getMeasureTable()->save($measure);
-					 		}
-					 	}
-					}
-				}
-				else {
-					$measure->exchangeArray($form->getData());
-					$this->getMeasureTable()->save($measure);
-				}
-			return $this->redirect()->toRoute('admin/measure');
+				return $this->redirect()->toRoute('admin/measure');
 			}
 		}
 		return array(
-				'form' => $form,
-				'config' => $this->config
-		);
+			'form' => $form,
+			'config' => $this->config
+			);
 	}
 
 	public function editAction()
@@ -104,12 +68,12 @@ implements ConfigAwareInterface
 		$id = (int) $this->params()->fromRoute('id', 0);
 		if (!$id) {
 			return $this->redirect()->toRoute('admin/measure', array(
-					'action' => 'add'
-			));
+				'action' => 'add'
+				));
 		}
 		$measure = $this->getMeasureTable()->get($id);
 		$previousImage = $measure->getImage();
-			
+
 		$form = $this->getServiceLocator()->get("Admin\Form\MeasureForm");
 		
 		$form->get("id")->setValue($measure->getId());
@@ -121,87 +85,44 @@ implements ConfigAwareInterface
 
 		$request = $this->getRequest();
 		if ($request->isPost()) {
-				
-			$measure->getInputFilter()->get('image')->setRequired(false);
+
 			$form->setInputFilter($measure->getInputFilter());
-				
-			$nonFile = $request->getPost()->toArray();
-			$file    = $this->params()->fromFiles('image');
-			$data = array_merge(
-					$nonFile,
-					array('image'=> $file['name'])
-			);
-				
+			$data = $request->getPost()->toArray();
 			$form->setData($data);
 
 			if ($form->isValid()) {
 				
-				$measureData = $form->getData();
+				$measure->setSpecification($request->getPost('specification'));
+				$measure->setMeasureType($request->getPost('measure_type'));
+				$measure->setMeasureValue($request->getPost('measure_value'));
+				$measure->setMeaning($request->getPost('meaning'));
+				$measure->setGeneralInformation($request->getPost('general_information'));
 				
-				$size = new Size($this->config['file_characteristics']['image']['size']);
-				$extension= new Extension($this->config['file_characteristics']['image']['extension']);
-				$image = $form->get('image')->getValue() ;
+				$fileService = $this->getServiceLocator()->get('Admin\Service\FileService');
+				$fileService->setDestination($this->config['component']['measure']['image_path']);
+				$fileService->setSize($this->config['file_characteristics']['image']['size']);
+				$fileService->setExtension($this->config['file_characteristics']['image']['extension']);
 
-				if(!file_exists($this->config['component']['measure']['image_path']))
-					mkdir($this->config['component']['measure']['image_path']);
+				$image = $this->params()->fromFiles('image');
 
-				$newImage = empty($data['image']) ? $previousImage : $data['image'] ;
-
-				$measure->setSpecification($measureData["specification"]);
-				$measure->setMeasureType($measureData['measure_type']);
-				$measure->setMeasureValue($measureData["measure_value"]);
-				$measure->setImage($newImage);
-				$measure->setMeaning($measureData["meaning"]);
-				$measure->setGeneralInformation($measureData["general_information"]);
-				
-				if(!empty($image)){
-					unlink($this->config['component']['measure']['image_path']."/".$previousImage);
-
-					$adapter = new HttpTransfer();
-					$adapter->setValidators(array($size,$extension), $file['name']);
-
-					if (!$adapter->isValid()) {
-							
-						$dataError = $adapter->getMessages();
-
-						$error = array();
-						foreach($dataError as $key=>$row)
-							$error[] = $row;
-
-						$form->setMessages(array('image'=>$error));
-					}
-					else {
-							
-						$adapter->setDestination($this->config['component']['measure']['image_path']);
-
-						foreach ($adapter->getFileInfo() as $info) {
-							$fileName = time().'.'.pathinfo($info['name'], PATHINFO_EXTENSION) ;
-								
-							$adapter->addFilter('File\Rename',
-									array('target' => $adapter->getDestination().'/'.$fileName.'',
-											'overwrite' => true));
-
-							if ($adapter->receive($info['name'])) {
-								$measure->setImage($fileName);
-								$this->getMeasureTable()->save($measure);
-							}
-						}
-					}
-				}
-				else {
-					$measure->setImage($previousImage);
-					$this->getMeasureTable()->save($measure);
+				if(isset($image['name']) && !empty($image['name'])) {
+					$image = $fileService->copy($image);
+					$measure->setImage($image);
+					if(isset($previousImage) && !empty($previousImage))
+						unlink($this->config['component']['measure']['image_path']."/".$previousImage);
 				}
 
+				$this->getBrandTable()->save($measure);
 				return $this->redirect()->toRoute('admin/measure');
+
 			}
 		}
-
 		return array(
-				'id' => $id,
-				'form' => $form,
-				'config' => $this->config
-		);
+			'id' => $id,
+			'image' => $previousImage,
+			'form' => $form,
+			'config' => $this->config
+			);
 	}
 
 	public function deleteAction()
@@ -223,10 +144,10 @@ implements ConfigAwareInterface
 			return $this->redirect()->toRoute('admin/measure');
 		}
 		return array(
-				'id'=> $id,
-				'config' => $this->config,
-				'measure' => $this->getMeasureTable()->get($id)
-		);
+			'id'=> $id,
+			'config' => $this->config,
+			'measure' => $this->getMeasureTable()->get($id)
+			);
 	}
 
 	public function setConfig($config)
